+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog
 
   Thanks to Michael Howitz in `PR #253 <https://github.com/adamchainz/django-upgrade/pull/253>`__.
 
+* Add Django 4.2+ fixer to combine deprecated ``DEFAULT_FILE_STORAGE`` and ``STATICFILES_STORAGE`` settings into the new ``STORAGES`` setting.
+
 * Add Django 4.0 fixer to rewrite ``django.contrib.admin.utils.lookup_needs_distinct`` -> ``lookup_spawns_duplicates``.
 
   Thanks to Bruno Alla in `PR #313 <https://github.com/adamchainz/django-upgrade/pull/313>`__.